// jshandymanservicesoftallyinc-from/fromt/src/app/models/product.model.ts
export class Product {

     idProduct: number;
     idBill: number;

     productCode: string;
     // productcode: string;
     name: string;
     categories: string;
     description: string;
     notas: string;
     price: number;
     priceWithTaxes: number;

     stockNumber: number;
     // stocknumber
     totalPriceWithTaxes: number;
     totalPriceWithOupTaxes: number;
     itemDeliteEdit: boolean;
}